feat : 유저 정보 초기화 hook 분리
